- refactor : api 폴더 종류별 분류
